+++ SimpleDocbase/Model/SettingItem.swift
@@ -1,0 +1,19 @@
+//
+//  SettingItem.swift
+//  SimpleDocbase
+//
+//  Created by jeonsangjun on 2017/10/30.
+//  Copyright © 2017年 jeon sangjun. All rights reserved.
+//
+
+import Foundation
+
+class SettingItem {
+    let name: String
+    let item: String
+    
+    init(name: String, item: String) {
+        self.name = name
+        self.item = item
+    }
+}
